Procesamiento de coordenadas optimizado
