+++ Hackathon/Streaming/ElasticWritter_Traffic.py
@@ -12,7 +12,7 @@ from elasticsearch import Elasticsearch
 import json
 import ast
 from pyproj import Proj, transform
-
+import numpy as np
 
 
 
@@ -31,11 +31,15 @@ class processTraficData(beam.DoFn):
         #Pasamos de string a array de coordenadas (float)
         coordenadas = []
         coord = ast.literal_eval(item['coordinates'])
-        #Bucle para conversión de coordenadas
-        for i in coord:
-            x2,y2 = transform(inProj, outProj,i[0],i[1])
-            c = [y2,x2]
+        
+        #Conversión de coordenadas
+        coords = np.asarray(coord)
+        x2,y2 = transform(inProj, outProj,coords[:,0],coords[:,1])
+        
+        for i,j in zip(x2,y2):
+            c = [j,i]
             coordenadas.append(c)
+        
         
         if item['idtramo']=='':
             item['idtramo']=0
@@ -48,8 +52,6 @@ class processTraficData(beam.DoFn):
                  'denominacion':item['denominacion'],
                  'modified':item['modified'],
                  'estado':int(item['estado']),
-                 'coordinates':item['coordinates'],
-                 'uri':item['uri'],
                  'location':{
                          "type":"linestring",
                          "coordinates":coordenadas
@@ -66,110 +68,7 @@ class IndexTraficDocument(beam.DoFn):
         res = self.es.index(index='prueba_gs3',body=element)
         
         print(res)
- 
-
-class processValenbisiData(beam.DoFn):
-    """
-    Filter data for inserts
-    """
-
-    def process(self, element):
-        
-        #{"empty_slots":17,"extra":{"address":"Economista Gay - Constituci\xc3\xb3n","banking":false,"bonus":false,"last_update":1578482815000,"slots":20,"status":"OPEN","uid":136},"free_bikes":3,"id":"1f6b81722ca23ce520f77207b868afa9","latitude":39.4899091610835,"longitude":-0.375701108044157,"name":"136_CALLE_ECONOMISTA_GAY","timestamp":"2020-01-08T11:34:20.782000Z"}'
-        
-        item = json.loads(element)
-        return [{'empty_slots':item['empty_slots'],
-                 'free_bikes':item['free_bikes'],
-                 'id':item['id'],
-                 'name':item['name'],
-                 'timestamp':item['timestamp'],
-                 'latitude':item['latitude'],
-                 'longitude':item['longitude'],
-                 'address':item['extra']['address'],
-                 'banking':item['extra']['banking'],
-                 'bonus':item['extra']['bonus'],
-                 'status':item['extra']['status'],
-                 'uid':item['extra']['uid'],
-                 'location':str(item['latitude'])+","+str(item['longitude'])                
-                 }]
-
-
-class IndexValenbisiDocument(beam.DoFn):
-   
-    es=Elasticsearch([{'host':'localhost','port':9200}])
     
-    def process(self,element):
-        
-        res = self.es.index(index='valenbisi',body=element)
-        
-        print(res)
- 
-        
-    
-def run_valenbisi(argv=None, save_main_session=True):
-  """Main entry point; defines and runs the wordcount pipeline."""
-  valenbisi_parser = argparse.ArgumentParser()
-  
-  #1 Replace your hackathon-edem with your project id 
-  valenbisi_parser.add_argument('--input_topic',
-                      dest='input_topic',
-                      #1 Add your project Id and topic name you created
-                      # Example projects/versatile-gist-251107/topics/iexCloud',
-                      default='projects/hackathon-diego/topics/valenbisi',
-                      help='Input file to process.')
-  #2 Replace your hackathon-edem with your project id 
-  valenbisi_parser.add_argument('--input_subscription',
-                      dest='input_subscription',
-                      #3 Add your project Id and Subscription you created you created
-                      # Example projects/versatile-gist-251107/subscriptions/quotesConsumer',
-                      default='projects/hackathon-diego/subscriptions/streaming',
-                      help='Input Subscription')
-  
-    
-  known_args, pipeline_args = valenbisi_parser.parse_known_args(argv)
-
-  # We use the save_main_session option because one or more DoFn's in this
-  # workflow rely on global context (e.g., a module imported at module level).
-  pipeline_options = PipelineOptions(pipeline_args)
-   
-  google_cloud_options = pipeline_options.view_as(GoogleCloudOptions)
-  #3 Replace your hackathon-edem with your project id 
-  google_cloud_options.project = 'hackathon-diego'
-  google_cloud_options.job_name = 'myjob'
- 
-  # Uncomment below and add your bucket if you want to execute on Dataflow
-  #google_cloud_options.staging_location = 'gs://edem-bucket-roberto/binaries'
-  #google_cloud_options.temp_location = 'gs://edem-bucket-roberto/temp'
-
-  pipeline_options.view_as(StandardOptions).runner = 'DirectRunner'
-  #pipeline_options.view_as(StandardOptions).runner = 'DataflowRunner'
-  pipeline_options.view_as(StandardOptions).streaming = True
-
- 
-  pipeline_options.view_as(SetupOptions).save_main_session = save_main_session
-
-
- 
-
-  p = beam.Pipeline(options=pipeline_options)
-
-
-  # Read the pubsub messages into a PCollection.
-  biciStations = p | beam.io.ReadFromPubSub(subscription=known_args.input_subscription)
-
-  # Print messages received
- 
-  
-  
-  biciStations = ( biciStations | beam.ParDo(processValenbisiData()))
-  
-  biciStations | 'Print Quote' >> beam.Map(print)
-  
-  # Store messages on elastic
-  biciStations | 'Bici Stations Stored' >> beam.ParDo(IndexValenbisiDocument())
-  
-  result = p.run()
-  #result.wait_until_finish()
 
 def run_trafic(argv=None, save_main_session=True):
   #Trafic data
@@ -231,10 +130,8 @@ def run_trafic(argv=None, save_main_session=True):
   traficstatus | 'Bici Stations Stored' >> beam.ParDo(IndexTraficDocument())
   
   result = p2.run()
-  #result.wait_until_finish()
+  result.wait_until_finish()
   
 if __name__ == '__main__':
   logging.getLogger().setLevel(logging.INFO)
   run_trafic()
-  run_valenbisi()
-
